[Refactor] #433 - LoginViewModel에서 Firebase에 직접 접근하는 코드를 AuthManager로 이동
